Rearranged timeline and added alert timer
Facility icons and timestamps have been placed on the timeline item rather than
to the left, and the items have generally been condensed.
The capturing faction's color is now used as a background as well.

On the map - the badge icons have been converted to view width dependent sizes
so that they don't crowd the map as much on small screens
Some minor considerations have been added for mobile, but its definitely not a
perfect experience on mobile.

A timer has been added to the top right corner of the map. It is removed
upon alert completion.

When zooming in and out the in game sounds are played. The alert start/end
sounds are present in the files but we are not using them yet.